ForceLock: new subcommand naming, add lock-force-release-client
The force-lock command name is ambigious with the addition of new
lock-handling subcommands.

Convert to a new naming scheme with the addition of a command to
forcibly release the lock for a single client.

This command is useful if a single client exits because it could not
acquire the global locks, and you only want to release that client's
lock to resume it.

Print a deprecation message for any old users of force-lock, and do not
include force-lock in the help output.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>